ai-service/requirements.txt
# ============================================================
# StyleLens V6 SOTA Pipeline — Requirements
# 7 SOTA Models + Hunyuan3D 2.0 3D Viewer
# Target: Apple M4 Pro 48GB (MPS)
# ============================================================

# Core
fastapi>=0.115
uvicorn[standard]>=0.34
python-multipart>=0.0.18
python-dotenv>=1.0

# ML / Vision
torch>=2.6
torchvision>=0.21
ultralytics>=8.4              # YOLO26-L (NMS-free person detection)
transformers>=4.48            # SAM 3D Body DINOv3
safetensors>=0.5
accelerate>=1.3
gguf>=0.4                     # FLUX.1-dev GGUF Q8 loader

# CatVTON-FLUX (FLUX-based virtual try-on)
diffusers>=0.32
peft>=0.15

# Image / 3D
opencv-python-headless>=4.10
Pillow>=11.0
numpy>=2.2
trimesh>=4.6
pygltflib>=1.16

# Gemini
google-genai>=1.5

# Face (optional — for face identity preservation)
insightface>=0.7
onnxruntime>=1.20

# Utils
aiohttp>=3.11
beautifulsoup4>=4.12
